requirements.txt
discogs_client==2.3.0
Flask==2.3.2
flask_wtf==1.2.2
python-dotenv==1.1.1
Requests==2.32.5
SQLAlchemy==2.0.25
WTForms==3.0.1
